split message functions off main.go. slight mod to tucoduel to specify username
