+++ go.mod
@@ -2,7 +2,7 @@ module github.com/fadedpez/tucoramirez
 
 go 1.20
 
-require github.com/bwmarrin/discordgo v0.27.1
+require github.com/bwmarrin/discordgo v0.28.1
 
 require (
 	github.com/gorilla/websocket v1.4.2 // indirect
