+++ main.go
@@ -1,16 +1,13 @@
 package main
 
 import (
-	"bufio"
 	"flag"
 	"fmt"
 	"github.com/bwmarrin/discordgo"
+	"github.com/fadedpez/tucoramirez/tucobot"
 	"log"
-	"math/rand"
 	"os"
 	"os/signal"
-	"regexp"
-	"strconv"
 	"syscall"
 	"time"
 )
@@ -37,7 +34,9 @@ func main() {
 		log.Printf("Logged in as: %s\n", s.State.User.Username)
 	})
 
-	dg.AddHandler(messageCreate)
+	// dg.AddHandler(messageCreate) // This is the old way of doing things
+
+	dg.AddHandler(tucobot.MessageCreate)
 
 	dg.Identify.Intents |= discordgo.IntentsGuildMembers
 	dg.Identify.Intents |= discordgo.IntentsGuildMessageReactions
@@ -69,63 +68,4 @@ func main() {
 		}
 		time.Sleep(1 * time.Second)
 	}
-}
-
-func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-	if m.Author.ID == s.State.User.ID {
-		return
-	}
-
-	if m.Content == "tucosay" {
-		_, _ = s.ChannelMessageSend(m.ChannelID, randFromTxt("quotes.txt"))
-	}
-
-	if regexp.MustCompile(`[tT][hH][aA][nN][kK][sS] [tT][uU][cC][oO]`).MatchString(m.Content) {
-		_, _ = s.ChannelMessageSend(m.ChannelID, "De nada, amigo.")
-	}
-
-	if m.Content == "tucoduel" {
-		tucoRoll := diceRoll()
-		userRoll := diceRoll()
-
-		tucoString := strconv.Itoa(tucoRoll)
-		userString := strconv.Itoa(userRoll)
-
-		if tucoRoll > userRoll {
-			_, _ = s.ChannelMessageSend(m.ChannelID, "Hurrah! Come back when you learn how to shoot cabrón! (Tuco: "+tucoString+" ; User: "+userString+")")
-		} else if tucoRoll < userRoll {
-			_, _ = s.ChannelMessageSend(m.ChannelID, "You pig! You haven't seen the last of Tuco! (Tuco: "+tucoString+" ; User: "+userString+")")
-		} else {
-			_, _ = s.ChannelMessageSend(m.ChannelID, "It seems we live to fight another day, friend. (Tuco: "+tucoString+" ; User: "+userString+")")
-		}
-	}
-
-	if regexp.MustCompile(`tuco\?$`).MatchString(m.Content) {
-		_, _ = s.ChannelMessageSend(m.ChannelID, randFromTxt("images.txt"))
-	}
-
-}
-
-func randFromTxt(path string) string {
-	file, err := os.Open(path)
-	if err != nil {
-		return ""
-	}
-	defer file.Close()
-
-	var quotes []string
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		quotes = append(quotes, scanner.Text())
-	}
-
-	quote := quotes[rand.Intn(len(quotes))]
-	return quote
-}
-
-func diceRoll() int {
-	min := 1
-	max := 100
-	r := rand.Intn(max-min) + min
-	return r
 }
